So I refined the pattern match at Specificity. It was not taking into account the possibility that the given period was not a supported one. Say Week, Month, or Year. It could be MILLENIA! (Yes, that is a valid enum from java.time.temporal.ChronoUnit. )
So, something had to be done, and something was done. Options were used, and now the library may or may not give you what you want. So, there is that. Now there was the question, should this uncertainty travel back all the way to the tests?

Should the tests take into account that TimeExpression may or may not return a value? Because of Specificity?

And the answer to that was a strong negative from my part, because unlike the library, I DO know the client of the library.
The library cannot trust the client, and so it is ready to respond that it just wont perform what the client wants it to, if the case seems to be that the client brought a MILLENIA to play with the library, which only supports Weeks, Months and Years.

So there is that. But like I said, I DO know the client, I made it myself. And I know for a fact that I will play fair with the library, and so, no, I will now allow the uncertainty to pass trought. That is my final answer. I believe we can stop it on its tracks, as fast as it comes out from the library.

So. Options were used. But only inside the library. The client inmediately goes for the value without hesitation, as it knows that there is no change of the library not sending anything other than the value it is looking for.